+++ apps/portal/app/buckets/[...parts]/_components/object-list-item.tsx
