+++ apps/portal/package.json
@@ -28,7 +28,6 @@
     "react-dom": "^19.0.0",
     "react-dropzone": "^14.3.5",
     "react-hook-form": "^7.54.2",
-    "react-intersection-observer": "^9.15.1",
     "react-time-ago": "^7.3.3",
     "viem": "2.x",
     "wagmi": "^2.14.8",
